Fix #4253: BubbleChartSeries switch x,y,radius to double from int.

+++ src/main/java/org/primefaces/model/chart/BubbleChartSeries.java
@@ -21,9 +21,9 @@ public class BubbleChartSeries implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-    private int x;
-    private int y;
-    private int radius;
+    private double x;
+    private double y;
+    private double radius;
     private String label;
 
     public BubbleChartSeries() {
@@ -33,7 +33,7 @@ public class BubbleChartSeries implements Serializable {
         this.label = label;
     }
 
-    public BubbleChartSeries(String label, int x, int y, int radius) {
+    public BubbleChartSeries(String label, double x, double y, double radius) {
         this.label = label;
         this.x = x;
         this.y = y;
@@ -48,27 +48,27 @@ public class BubbleChartSeries implements Serializable {
         this.label = label;
     }
 
-    public int getRadius() {
+    public double getRadius() {
         return radius;
     }
 
-    public void setRadius(int radius) {
+    public void setRadius(double radius) {
         this.radius = radius;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 }
